Minor cleanup before adding libshout stuff
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@22806 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: dbd8ae653459296aefc93f7801304e52b21f49c8

+++ sandbox/compressed_video_streaming/image_decompression_node.cpp
@@ -149,7 +149,7 @@ public:
     {
     	if(oggpacket.packetno != 0)
     	{
-    		ROS_DEBUG("Dumping header packet because packet# is non-zero: %d.", oggpacket.packetno);
+    		ROS_DEBUG("Dumping header packet because packet# is non-zero: %d.", (int)oggpacket.packetno);
     		return;
     	}
     	//static th_setup_info* setup_info_ptr = null;
